tidy up tests for /corrections
